requirements.txt
django
pillow
djangorestframework
requests
django-dynamic-scraper
scrapy
django-celery